+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- AES/GCM: Custom key size support `AesGcmEncryption(x)` where `x` is the valid key size in bits.
+    - Valid key sizes: `128`, `192`, `256`.
+    - Custom key sizes are **optional**, and will default to `256` if not specified.
+
 ### Changed
 
 ### Removed
@@ -32,7 +36,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Project structure relayout (composite builds)
 - Throw custom exception instead of generic
 - **BREAKING CHANGE:** reduce package nesting to `kiper.{module}`
-
 
 ## [0.1.0] - 2023-01-07